fix: remove namespace for AsyncAvatarStack
this is also a better naming which matches closer with things like AsyncImage.
